Crate resources

Manages system resources that can be allocated to VMs and their devices.

Modules

address_allocator

address_range

system_allocator

Structs

AddressRange

Represents a range of addresses from start to end, inclusive.

AllocOptions

Memory allocation options.

SystemAllocator

SystemAllocatorConfig

Enums

Alloc

Used to tag SystemAllocator allocations.

Error

MmioType

Manages allocating system resources such as address space and interrupt numbers. MMIO address Type Low: address allocated from low_address_space High: address allocated from high_address_space

Type Definitions

Result