Struct resources::AddressAllocator

pub struct AddressAllocator { /* fields omitted */ }

Manages allocating address ranges. Use AddressAllocator whenever an address range needs to be allocated to different users. Allocations must be uniquely tagged with an Alloc enum, which can be used for lookup. An human-readable tag String must also be provided for debugging / reference.

Examples

// Anon is used for brevity. Don't manually instantiate Anon allocs!
  AddressAllocator::new(0x1000, 0x10000, Some(0x100)).map(|mut pool| {
      assert_eq!(pool.allocate(0x110, Alloc::Anon(0), "caps".to_string()), Ok(0x1000));
      assert_eq!(pool.allocate(0x100, Alloc::Anon(1), "cache".to_string()), Ok(0x1200));
      assert_eq!(pool.allocate(0x100, Alloc::Anon(2), "etc".to_string()), Ok(0x1300));
      assert_eq!(pool.get(&Alloc::Anon(1)), Some(&(0x1200, 0x100, "cache".to_string())));
  });

Implementations

impl AddressAllocator

pub fn new(
pool_base: u64,
pool_size: u64,
align_size: Option<u64>
) -> Result<Self>

Creates a new AddressAllocator for managing a range of addresses. Can return None if pool_base + pool_size overflows a u64 or if alignment isn’t a power of two.

pool_base - The starting address of the range to manage.

pool_size - The size of the address range in bytes.

align_size - The minimum size of an address region to align to, defaults to four.

pub fn pool_base(&self) -> u64

Gets the starting address of the allocator.

This returns the original pool_base value provided to AddressAllocator::new().

pub fn pool_size(&self) -> u64

Gets the size of the allocator’s address range in bytes.

This returns the original pool_size value provided to AddressAllocator::new().

pub fn reverse_allocate_with_align(
&mut self,
size: u64,
alloc: Alloc,
tag: String,
alignment: u64
) -> Result<u64>

Allocates a range of addresses from the reverse managed region with an optional tag and minimal alignment. Returns allocated_address. (allocated_address, size, tag) can be retrieved through the get method.

pub fn allocate_with_align(
&mut self,
size: u64,
alloc: Alloc,
tag: String,
alignment: u64
) -> Result<u64>

Allocates a range of addresses from the managed region with an optional tag and minimal alignment. Returns allocated_address. (allocated_address, size, tag) can be retrieved through the get method.

pub fn allocate(&mut self, size: u64, alloc: Alloc, tag: String) -> Result<u64>

pub fn allocate_at(
&mut self,
start: u64,
size: u64,
alloc: Alloc,
tag: String
) -> Result<()>

Allocates a range of addresses from the managed region with an optional tag and required location. Allocation alignment is not enforced. Returns OutOfSpace if requested range is not available (e.g. already allocated with a different alloc tag).

pub fn release(&mut self, alloc: Alloc) -> Result<()>

Releases exising allocation back to free pool.

pub fn release_containing(&mut self, value: u64) -> Result<()>

Release a allocation contains the value.

pub fn get(&self, alloc: &Alloc) -> Option<&(u64, u64, String)>

Returns allocation associated with alloc, or None if no such allocation exists.

pub fn address_from_pci_offset(
&self,
alloc: Alloc,
offset: u64,
size: u64
) -> Result<u64>

Returns an address from associated PCI alloc given an allocation offset and size.